Moved some tests to test_fields_from.py
